Implemented changes to support JWT
