+++ server-users/package.json
@@ -10,6 +10,8 @@
     "dependencies": {
         "bcrypt": "5.1.1",
         "express": "4.18.2",
-        "mysql2": "2.3.3"
+        "mysql2": "2.3.3",
+        "crypto": "1.0.1",
+        "jsonwebtoken": "9.0.2"
     }
 }
